backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
opencv-python==4.8.1.78
tensorflow==2.14.0
torch==2.1.1
torchvision==0.16.1
numpy==1.26.2
pandas==2.1.3
pymongo==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
kafka-python==2.0.2
pillow==10.1.0
python-socketio==5.10.0
aiofiles==23.2.1
motor==3.3.1
sqlalchemy==2.0.23
alembic==1.12.1
pyjwt==2.8.0
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cublas-cu12==12.1.3.1
nvidia-cudnn-cu12==8.9.2.26
